Fix Straight Line Bug
In some case, if use draw a nodule area as a straight line align to
x-axis or y-axis. The area will be 0. And this casue a bug which make
program crash.
So in this update, we fixed this. If a new nodule is created and the
area is 0, the width or length will be expend to 1 to make area > 0.

Change the 'checked status' button color to make it more noticable.